Implement error contexts for MBS (#1192)
### What

This PR allows MBS errors to provide a larger "context". Our notion of
context is an ordered list of annotated paths. An example would be for a
name conflict, where we want to say `["this name was used here", "...
and also here"]`, so reporting two paths would be helpful.

Note that the actual format of `path` is a structured data type, not a
string. An example of the first instance introduced by this PR can be
seen here:

```json
{
  "errors": [
    {
      "context": [
        {
          "message": "Data connector name given here",
          "path": [
            { "Index": 11 },
            { "Key": "definition" },
            { "Key": "source" },
            { "Key": "dataConnectorName" }
          ]
        }
      ],
      "code": "opendds-validation",
      "message": "invalid metadata: error building schema: invalid metadata: the source data connector b (in subgraph default) for model Artists (in subgraph default) has not been defined",
      "path": null
    }
  ]
}
```

The `path` key is already generated by other parts of the MBS process,
so we preserve it for now to avoid breaking changes.

### How

* A new crate, `error-context`, to hold the type of the context
structure.
* Otherwise essentially the same as the path mechanism.

V3_GIT_ORIGIN_REV_ID: 0bb529542ba0a7f6774c08fa6adb7d7c608d6c5b