Use `BTreeSet` to collect `Role` in `metadata-resolve` (#1312)
<!-- The PR description should answer 2 important questions: -->

### What

JSONAPI builds iterate over `roles` in metadata, and one was very slow,
realised it was because it was calculating for `admin` over and over.
This changes the `roles` in `metadata-resolve` to be collected in a
`BTreeSet` to remove duplicates, which fixes the problem.

### How

Change `Vec` to `BTreeSet`, fix type errors and snapshots.

V3_GIT_ORIGIN_REV_ID: c61ef9e490c3cde9d004c08ef71dc73e92cd6e7e