`OperatorMapping` struct newtype (#1336)
<!-- The PR description should answer 2 important questions: -->

### What

Operator mappings in boolean expressions allow us to override the name
of an operator, so we can call some database's `_eq` operator the much
more attractive `equals`. If not override is provided we can just use
the database's name directly. We make this a newtype with a `get`
function to encode this logic and make doing the right thing the easiest
thing.

V3_GIT_ORIGIN_REV_ID: e47f4e291494562ed1551d0d7469259198862a0a