Make the docker-compose usable with engine running locally (#1372)
### What
This PR makes the docker-compose.yaml useful for both running the engine
in docker as well as when the engine is running locally on the host.
This enables a new `just start-engine` target that runs the engine
locally while using the same deps and configuration as in
docker-compose.yaml. Running engine this way is faster than constantly
rebuilding in Docker.

### How
The various component in docker-compose.yaml are configured to talk to
one another via `local.hasura.dev` in the same way as Hasura projects
are done via `ddn`. This enables the engine to not care whether it is
running on the host machine directly via cargo or inside a docker
container.

The Dockerfile has also been simplified a little and now only builds the
engine, not every single crate.

V3_GIT_ORIGIN_REV_ID: b17b066fc204fd76b74633990b9862f5db4f180a