Copy of `plan` generation moved to `graphql-ir` (#1380)
<!-- The PR description should answer 2 important questions: -->

### What

The `execute` crate doesn't need to know about generating
`QueryExecutionPlan` etc from GraphQL IR. We move the planning to
`graphql-ir`. It is mostly moved, this is a large enough set to merge to
keep the size down.

This creates a new pipeline that is only switched when a static value is
changed. This allows iterating on it without breaking the engine.

V3_GIT_ORIGIN_REV_ID: a45874c026778071be2300bc915954824618fe14