Remove `EngineRouter` abstraction (#1418)
<!-- The PR description should answer 2 important questions: -->

### What

We want to decompose the engine routes so we can split `sql` out, so
first we need to remove this abstraction, so instead the app is in
charge of manually adding the routes it wants.

Functional no-op.

V3_GIT_ORIGIN_REV_ID: 1e1ea4f3797d289c91299765a7285c423fef00a3