Remove panic in type mappings, allow arrays in command return types (#1425)
<!-- The PR description should answer 2 important questions: -->

### What

When an array return type was used in a command that we using in type
mappings we panicked, however this should be allowed. Return a proper
error if a Predicate is used as a command return type though, as that is
not allowed.

This fixes the immediate error in the ticket, but the reproducer throws
another one - I think this is because the reproduction is so minimal
though. Worth testing with the full reproduction before merging
@jberryman .

V3_GIT_ORIGIN_REV_ID: ae0a2f0a6fe83c82e5ab01adaba9b317743688ba