Refactored Flags into a Set and made false flags not serialize (#1424)
### What
OpenDD flags are now stored as a Set of a new Flag enum type. This
allows us to enumerate the flags easily (via the `IntoEnumIterator`
trait), which will be useful in the future to generate a list of
compatibility dates. The new OpenDdFlags set type deserializes to
exactly the same JSON as before, with one exception: any "false" flags
(ie. flags not in the set) are not serialized. This should help newer
MBS build metadata that is compatible with older engine versions that
don't understand newer flags that are disabled and therefore irrelevant.

### How
The main change is in `crates/open-dds/src/flags.rs` where a new
OpenDdFlags type has been created. It is just a wrapper around a
`BTreeSet` of our new `Flag` enum. It has a custom `Serialize`,
`Deserialize` and `JsonSchema` implementation that performs the
serialization we want, and also manually creates the same JsonSchema as
we previously had that matches the serialization.

Unit tests at the bottom of that file perform roundtrip tests and check
that the serialization does what we want.

Now that the flags are not a struct and are instead a set, all usages
change to using `.contains`. You can see this peppered through
metadata_resolve.

In MBS,
`crates/cloud/v3-metadata-build-service/src/build/compatibility/flags.rs`
is where we rework the compatibility date handling to leverage the enum.
We perform a pattern match on it to ensure there's a compatibility date
added for new flags and we now automatically compute the correct flags
to set based on the date using that match and iterating over the flags.

V3_GIT_ORIGIN_REV_ID: 284529db3f346e8eb635a47c9a8fa19d4b8ad207