[PQL-72] Expose models with arguments as table-valued functions (#1445)
### What

This adds models which have arguments to the SQL context, as
table-valued functions instead of regular tables like before (which was
an error).

<!-- Consider: do we need to put new checks behind a flag? -->

### How

Check each model during construction of the SQL context. If it has
arguments, add an `TableValuedFunction::ModelWithArguments` to the
commands list, instead of a table. Pattern match on
`TableValuedFunction` to determine what type of thing to register as a
user-defined TVF in `datafusion`.

V3_GIT_ORIGIN_REV_ID: 09951f3ba3e2549f57c4cbe5a4b8be96729e11d5