Support NDC 0.2.0 count aggregate types (#1451)
### What
This PR adds support for NDC 0.2.0's ability for data connectors to tell
the engine what scalar type count aggregations are returned in. Prior to
this PR, it was assumed every count returned the inbuilt Int type in
GraphQL and an unspecified JSON number type in NDC.

Now, for NDC 0.2.0 connectors, we need to look at the scalar type
specified as used by the count aggregations and ensure this type is used
in metadata. This means a new `returnType` field has been added to count
aggregates on `AggregateExpressions`:

```diff
kind: AggregateExpression
version: v1
definition:
  name: ManufacturersAggExp
  operand:
    object:
      aggregatedType: Manufacturers
      aggregatableFields:
        - fieldName: id
          aggregateExpression: UuidAggExp
        - fieldName: name
          aggregateExpression: TextAggExp
  count:
    enable: true
+   returnType: CustomInt
  countDistinct:
    enable: true
+   returnType: CustomInt
  graphql:
    selectTypeName: ManufacturersAggExp
```

The type used on this field must match what the connector declares. If
it is omitted, the inbuilt `Int` type is defaulted to (this is
effectively the existing behaviour). If the data connector does not
specify an count aggregate scalar type (ie. NDC 0.1.x), Int must be
specified (or defaulted to). Because of this, this is not a breaking
change for NDC 0.1.x connectors and metadata. However, for NDC 0.2.0
connectors, if they have a custom scalar as their representation for
their declared count aggregate type, then they will need to start using
this new return type property.

### How
`metadata_resolve` has been updated to now validate the new `returnType`
field in count definitions in OpenDD. It is first resolved in the
`aggregates` stage, but the chosen type is only checked against the data
connector during the `models` stage when the data connector is known.

In `graphql/schema` we use the resolved type in the schema, rather than
just hardcoding `Int`.

There are new failing `metadata_resolve` tests that test all the error
conditions, and the
`crates/metadata-resolve/tests/passing/aggregate_expressions/root_field/`
test has been updated to use NDC 0.2.0 and use a custom scalar.

A new test has been added to `engine` tests that uses the custom
connector to use a custom scalar for the count aggregates and see it
impact the GraphQL schema.

V3_GIT_ORIGIN_REV_ID: f01a98ae2749abb444645e1e0ca9c82fda9a8f6c