Use `ndc-spec 0.2.0` comparison operator meanings (#1433)
<!-- The PR description should answer 2 important questions: -->

### What

Store comparison operator meanings in `data_connector_scalar_types`.
These are used in the OpenDD pipeline so that SQL can ask for the
"greater than operator for Foo` etc.

Changed `OperatorName` in OpenDD pipeline so the special operators and
the custom operators are separated.

Previously we stored a `Vec` of operators, now we choose the first one
(which we were already doing in the case of duplication), raising a
warning if another is found. Said warning is raised to an error after a
compatibility config date.

V3_GIT_ORIGIN_REV_ID: 7c2bc98a1bafc60d08f814713244891c14b973aa