Filtering by nested scalar arrays (#1460)
### What

Consider an `ObjectType` with a field that is an array of scalars, such
as the `Country` type with its `cities` array of string field below.

```yaml
kind: ObjectType
version: v1
definition:
  name: Country
  fields:
    - name: id
      type: Int!
    - name: name
      type: String!
    - name: cities
      type: [String!]!
  graphql:
    typeName: Country
  dataConnectorTypeMapping: {}
```

It is now possible to write the following `where` clause in GraphQL:

```graphql
query {
  CountryMany(where: { cities: { _eq: "Melbourne" } }) {
    id
    name
    cities
  }
}
```

In order to do this, the `cities` field must be added as a
`comparableField` to Country's `BooleanExpressionType`:

```yaml
kind: BooleanExpressionType
version: v1
definition:
  name: Country_bool_exp
  operand:
    object:
      type: Country
      comparableFields:
        - fieldName: cities
          booleanExpressionType: String_bool_exp
  logicalOperators:
    enable: true
  isNull:
    enable: false
  graphql:
    typeName: Country_bool_exp
```

The `booleanExpressionType` of the `comparableField` must be for the
scalar type of the array element (in this case a `BooleanExpressionType`
for `String`).

However, doing this is only supported if the data connector your `Model`
sources from itself supports filtering by nested scalar arrays. This
support is currently in development and is coming soon (it is only
supported in NDC 0.2.0).

### How
In `metadata_resolve`, we now track a new ScalarComparisonKind that
captures a ScalarArray. We validate that the boolean expressions can
only be used on models that source from connectors that support the
appropriate capability.

The ScalarArray kind is passed through to GraphQL annotations as a new
ObjectFieldKind, which is consumed in GraphQL IR in order to know if the
field is a scalar array comparison. If so, it uses a new
`Expression::LocalNestedScalarArray` plan type to capture that.

This is eventually translated into the appropriate
`ExistsInCollection::NestedScalarCollection` in NDC v0.2.0. NDC v0.1.x
raises an error, but the lack of a capability will block the code from
producing `Expression::LocalNestedScalarArray`.

There are new passing and failing metadata_resolve tests, as well as a
new engine integration test. The custom connector has been updated with
support for scalar array filtering, and that is what is used in the
engine integration test.

V3_GIT_ORIGIN_REV_ID: a523bf395d4a4c8c7276dccabad9fb63433c2ed6