Introduce OpenDD pipeline behind flag (#1465)
<!-- The PR description should answer 2 important questions: -->

### What

This PR adds the boring plumbing to allow us to start using the OpenDD
pipeline for GraphQL. Functional no-op for users.

We add:
- an unstable feature `--unstable feature
enable-open-dd-pipeline-for-graphql` for single-tenant engine
- an enum in `ModelSelectMany` in `graphql-ir` that allows us to either
plan and execute with the new pipeline. Currently it explodes if used as
we need to modify the GraphQL -> OpenDD IR functions in
`graphql-frontend` and use those. This PR is big enough.
- we make `metadata_resolve::Metadata` available when planning a GraphQL
request so it can be used by `plan` crate

What's next?
- Change engine GraphQL tests so they run old / new pipeline and compare
results
- Remove old way of testing OpenDD pipeline
- Start making OpenDD and GraphQL work, starting with `ModelSelectMany`

V3_GIT_ORIGIN_REV_ID: 62ca59616230d1b84f5734b8a3af25cb18fc2ee7