Fix for input field presets on TypePermissions causing invalid typechecking errors at query time (#1502)
### What
This PR fixes a bug where erroneous typechecking errors could be
produced when querying a Model or Command that used an object type as an
argument that had input field presets applied on its TypePermissions.

This was because the type of the field that was recorded against the
preset was not the actual field type, but was incorrectly set to the
type of the argument/field that the object type came from. So
typechecking was being done against the wrong type.

### How

`crates/metadata-resolve/src/stages/argument_presets/presets.rs`
contains the real fix where we simply get the field's type and use it
instead of the incorrect type.

There's a new passing metadata_resolve test that captures the resolved
output in which we can see the correct type being saved in the preset,
and also an engine test has been update to use field presets to
demonstrate it working end to end.

V3_GIT_ORIGIN_REV_ID: 62efe9a9df2d8080990662aaf767c238c8fd671d