Partially fix argument and type field presets (#1523)
### What
This PR partially fixes argument and type field presets. Unfortunately
the whole approach needs to be reworked to fix all the issues, but these
fixes to the existing approach patch some of the most urgent issues.
Namely:

* If you define field presets on TypePermissions and that type is used
in a command/model argument, the other command/model argument presets
will get erased. This bug snowballs into a security hole because it
means that if you preset a pre/post_check boolexp on a pg mutation
command, that boolexp won't be sent to the connector because its been
erased, meaning that your permissions are not enforced.

* If you define field presets on TypePermissions for a parent object and
a some more on child of that parent object's TypePermissions, the parent
object presets will be erased

* If you define field presets on TypePermissions, and that type is used
in an argument but the argument type is an array of that object type,
the field presets will not be applied (for example, the `objects`
argument on pg mutations)

The following issues are still present but will require a rework of the
whole presetting system that will done in a follow-up PR:

* If you define field presets on TypePermissions and that type is used
nested inside an argument's type somewhere, but as an array of the
object type, the field presets will not be applied

* If you define field presets on TypePermissions and your types are
recursive, field presets will only be applied to the first layer and not
past the point where the types recur

### How

In `crates/metadata-resolve/src/stages/argument_presets/presets.rs`,
argument presets and field presets were computed and each _overwrote_
the current presets in the Role->ArgumentPresets map. This incredibly
wrong behaviour has been replaced with logic that updates the presets in
the Role map, meaning that field presets don't overwrite argument
presets and child field presets (as one recurs through the field preset
types) don't overwrite parent field presets.

In `crates/plan/src/query/arguments.rs` new handling has been added when
the argument is an array type. The existing logic is just looped over
the elements of the array rather than just not applying the presets.
This is not a full solution as it doesn't handle arrays nested in the
data structure, but it handles the most common use case for now and will
be replaced with a real solution soon.

In the metadata_resolve tests the
`crates/metadata-resolve/tests/passing/command_permissions/input_type_permissions`
test has been updated with various nested types with field presets to
ensure the presets are correctly captured in the final build metadata
snapshot.

A new `eval_institutions` function has been added to the custom
connector that takes an array of the Institution nested data structure
as an argument and a boolean expression to evaluate as another argument.
This allows a new execute test
(`crates/engine/tests/execute/commands/functions/boolean_expression_command_argument/combined_with_type_permissions`)
to test setting both a field preset into an array argument and an
argument preset at the same time.

V3_GIT_ORIGIN_REV_ID: b79c97e5d53c3fa93f33556c3500fdf9ab316737