add OpenAPI schema for filter in jsonapi (#1553)
### What

This PR adds the filter parameter in the OpenAPI schema for the JSON
API.

![image](https://github.com/user-attachments/assets/983a177b-54df-49bc-85fe-b86cc56ede45)

### How

We just iterate over the model's filter expression and build the schema.

*
[`crates/jsonapi/src/schema.rs`](diffhunk://#diff-05d0e381264b290f1d890504e9bcf08843e34f73233cd8d7aa4033e65db16184R64-R66):
Added logic to include a filter parameter if the model has a
`BooleanExpressionType`.
*
[`crates/jsonapi/src/schema/parameters.rs`](diffhunk://#diff-d6dbc8fb1cb650a070dc7f58a1d0093ac38e4777b15b6f94963647d537b11b34R237-R414):
Introduced `filter_parameters` function to generate a filter parameter
for models with `BooleanExpressionType`. This includes detailed schema
definitions for scalar, object, and relationship fields, as well as
logical operators like `$and` and `$or`.

V3_GIT_ORIGIN_REV_ID: 4563a8e27cfaaeab07563b57fe4ad06cf8923ad6