Update action-handlers.mdx
Using underscores in headers can cause issues with nginx.  Given the number of  comments on this stackoverflow:

https://stackoverflow.com/questions/22856136/why-do-http-servers-forbid-underscores-in-http-header-names

And nginx being the a very common production server:
https://nginx.org/en/docs/http/ngx_http_core_module.html#underscores_in_headers

I recommend you update the docs to use dashes instead of underscores.  While underscores are technically valid according to RFC, you are causing undo pain on devs finding this querk.