Gardening: Remove unnecessary Arc type usages (#1586)
### What
This gardening PR removes a bunch of usages of the Arc type which are
unnecessary. Unless code is going to clone the Arc to store it somewhere
(or needs to use some other Arc feature), it should simply take a
standard reference to the Arc-owned type.

V3_GIT_ORIGIN_REV_ID: 0c7137fb7c3e4366617f1ffcd23913cccb377a63