Fix AuthConfig JWT claimsConfig locations deserialization discarding errors and JSON schema issues (#1604)
### What
This PR fixes an issue where invalid JSON pointers specified inside
AuthConfig's `jwt.claimsConfig.locations` would cause all extra claims
added there to be silently dropped.

It also fixes the AuthConfig JSON schema generation so that JWTClaimsMap
actually specifies the proper type of the additional properties allowed
on that object (which are the custom claims). The lack of this explicit
definition is likely why the CLI is dropping these claims when it
creates the metadata to send to MBS.

### How
The silent dropping of the extra claims when an invalid JSON pointer was
encountered is because of the weird interaction between serde's flatten
and Option. In this case, a deserialization error causes flatten to
throw away the error and just return a None for the option. This has
been fixed by removing the Option which was unnecessary anyway.

This change introduces "new" errors, but existing builds won't have any
custom claims anyway, since the CLI has been dropping them, so we've
dodged a bullet and don't need a flag and a workaround to deal with
that. Schema diff tests are still passing, which appears to validate
this hypothesis.

The JWTClaimsMap JSON schema issue has been worked around with an extra
macro attribute.

V3_GIT_ORIGIN_REV_ID: 2c6b53be6d567ec5173737097625727040afe9be