Fix queries with filters that includes remote relationships from nested fields (#1605)
<!-- The PR description should answer 2 important questions: -->

### What

Queries with filter expressions that include remote relationships
applied from nested fields were failing. This error arises from
incorrectly assigning the leaf node of a field path as the source column
when the root column in the path should be considered as the source
column.

### How

Treat the first column in the field_path as a source column and append
current column to the nested field path. It is in consistent with the
logic of resolving local field comparison expression.

<!-- How is it trying to accomplish it (what are the implementation
steps)? -->

V3_GIT_ORIGIN_REV_ID: 8abeb1ce964526762524b58b80c05d76b7b657d4