ENG-1438: support remote joins from Procedures (#1628)
This was supported in metadata, but failed at query time as the
execution component was still TODO. So this is a bug fix from the user's
point of view.

We also do a little refactoring for consistency with the Query code
path.

<!-- The PR description should answer 2 important questions: -->

V3_GIT_ORIGIN_REV_ID: 702bc60dc1faffa333f89d1f46a6ec7e8420ffa2