OpenDD `BooleanExpression` relationship change (#1674)
<!-- The PR description should answer 2 important questions: -->

### What

This PR changes the `open-dds::query::BooleanExpression` to have a
`Relationship` arm. Previously, all nesting happened within the
`Operand` on the `Comparison` arm, however this meant we could not
express the different between:

-----
```
And [
  RelationshipFoo (
   ("type" == "dog")
   ),
   RelationshipFoo (
     ("hasTail" == true)
   )
]
```

ie, "match when there is a field in `RelationshipFoo` where `type` ==
"dog" and another field in `RelationshipFoo` where `hasTail` == true"

----

```
RelationshipFoo (
  And [
   ("type" == "dog"),
   ("hasTail" == true)
  ]
)
```

ie, "match when there is a field in `RelationshipFoo` where `type` ==
"dog" AND `hasTail` == true"

This PR makes this change so we use this instead, and fixes a bunch of
tests in the process.

V3_GIT_ORIGIN_REV_ID: 21b770400b1eb0095a4cc528219405079e2ec72d